// kodikosERP-master/src/app/models/formatki/contractorform.model.ts
export interface Contractorform {
  balance?: boolean;
  comments?: boolean;
  creditLimit?: boolean;
  paymentDeadline?: boolean;
  status?: boolean;
  buttonDodajKontr?: boolean;
  buttonOfertaKontr?: boolean;
  buttonEdytujKontr?: boolean;
  buttonUsunKontr?: boolean;
  kontakty?: boolean;
  kontBankowe?: boolean;
  projekty?: boolean;
  firmy?: boolean;
  widziWszystkie?: boolean;
  adres?: boolean;
  pliki?: boolean;
  buttonPliki?: boolean;
}